move simplified AddressMap to system_components
